pyproject.toml
[project]
name = "binance-quant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "python-binance",
  "pandas",
  "matplotlib",
  "scikit-learn",
  "numpy",
  "torch",
  "pyarrow"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"